llm-server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
av==14.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.6.0
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
jinja2==3.1.6
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
numpy==2.2.5
ollama==0.4.8
onnxruntime==1.21.1
packaging==25.0
protobuf==6.30.2
pydantic==2.11.4
pydantic-core==2.33.2
pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.5
setuptools==80.3.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
tqdm==4.67.1
typer==0.15.3
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
openai==1.78.1